requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-community
python-dotenv
psycopg2-binary
fastapi
uvicorn
pytest
hypothesis